include hierarchy for native types fix for clang
