+++ elastic/rational/decimal/arithmetic/division.cpp
@@ -39,7 +39,8 @@ try {
 #if MANUAL_TESTING
 
 	using Rational = sw::universal::erational;
-
+	Rational r{1.2345};
+	std::cout << r << '\n';
 
 	ReportTestSuiteResults(test_suite, nrOfFailedTestCases);
 	return EXIT_SUCCESS; // ignore failures
