+++ include/universal/common/number_traits_reports.hpp
@@ -1,7 +1,7 @@
 #pragma once
 // number_traits.hpp: utility functions for working with numeric_limits<> 
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
