+++ include/universal/native/manipulators.hpp
@@ -15,70 +15,6 @@
 
 namespace sw { namespace universal {
 
-	template<typename IntegralType, 
-		std::enable_if_t< ::std::is_integral<IntegralType>::value, bool> = true
-	>
-	std::string type_tag(IntegralType = {}) {
-		// can't use a simple typeid(Real).name() because gcc and clang obfuscate the native types
-		constexpr unsigned nbits = sizeof(IntegralType) * 8;
-		std::string type_string;
-		if constexpr (nbits == 8) {
-			type_string = ::std::is_signed<IntegralType>::value ? std::string("int8_t") : std::string("uint8_t");
-		}
-		else if constexpr (nbits == 16) {
-			type_string = ::std::is_signed<IntegralType>::value ? std::string("int16_t") : std::string("uint16_t");
-		}
-		else if constexpr (nbits == 32) {
-			type_string = ::std::is_signed<IntegralType>::value ? std::string("int32_t") : std::string("uint32_t");
-		}
-		else if constexpr (nbits == 64) {
-			type_string = ::std::is_signed<IntegralType>::value ? std::string("int64_t") : std::string("uint64_t");
-		}
-		else {
-			type_string = std::string("unknown");
-		}
-		return type_string;
-	}
-
-	template<typename RealType,
-		std::enable_if_t< ::std::is_floating_point<RealType>::value, bool> = true
-	>
-	std::string type_tag(RealType = {}) {
-		// can't use a simple typeid(Real).name() because gcc and clang obfuscate the native types
-		constexpr unsigned nbits = sizeof(RealType) * 8;
-		std::string type_string;
-		if constexpr (nbits == 32) {
-			type_string = std::string("float");
-		}
-		else if constexpr (nbits == 64) {
-			type_string = std::string("double");
-		}
-		else if constexpr (nbits == 128) {
-			type_string = std::string("long double");
-		}
-		else {
-			type_string = std::string("unknown");
-		}
-		return type_string;
-	}
-
-	template<typename RealType,
-		std::enable_if_t< ::std::is_floating_point<RealType>::value, bool> = true
-	>
-	std::string type_field(RealType = {}) {
-		std::string fields{};
-		if constexpr (sizeof(RealType) == 4) {
-			fields = "fields(s:1|e:8|f:23)";
-		}
-		else if constexpr (sizeof(RealType) == 8) {
-			fields = "fields(s:1|e:11|f:52)";
-		}
-		if constexpr (sizeof(RealType) == 16) {
-			fields = "fields(s:1|e:15|f:112)";
-		}
-		return fields;
-	}
-
 	// internal function to extract exponent
 	template<typename Uint, typename Real>
 	int _extractExponent(Real v) {
@@ -161,6 +97,7 @@ namespace sw { namespace universal {
 		std::cout << "color  : " << color_print(value) << '\n';
 	}
 
+#ifdef DEPRECATED
 	// generate a binary string for a native IEEE floating point
 	template<typename Real,
 		typename = typename std::enable_if< std::is_floating_point<Real>::value, Real >::type
@@ -200,6 +137,7 @@ namespace sw { namespace universal {
 
 			return s.str();
 	}
+#endif
 
 	// return in triple form (sign, scale, fraction)
 	template<typename Real,
