+++ tools/cmd/single.cpp
@@ -1,20 +1,15 @@
 // single.cpp: components of a float: cli to show the sign/scale/fraction components of a single precision float 
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/utility/directives.hpp>
-#include <universal/utility/bit_cast.hpp>
-#include <limits>
-#define BITBLOCK_THROW_ARITHMETIC_EXCEPTION 0
-#define VALUE_THROW_ARITHMETIC_EXCEPTION 0
-#include <universal/native/ieee754.hpp>
-#include <universal/number/cfloat/cfloat.hpp>
 #include <universal/common/number_traits_reports.hpp>
+#include <universal/native/ieee754.hpp>
 
 // ShowRepresentations prints the different output formats for the float type
 template<typename Scalar>
-void ShowRepresentations_(std::ostream& ostr, float f) {
+void ShowRepresentations(std::ostream& ostr, float f) {
 	using namespace sw::universal;
 	auto oldprec = ostr.precision(); // save stream state
